requirements.txt
PyMySQL==0.7.9
mysql-replication==0.10